requirements.txt
aiogram
aiohttp
python-dotenv
aiofiles
mutagen
httpx
loguru
pillow
pydub
tempfile